画像(png)での支出登録機能を追加

+++ src/main/java/com/example/webapp_householdaccounts/entity/Expense.java
@@ -5,7 +5,7 @@ import jakarta.persistence.*;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.sql.Date;
+import java.util.Date;
 
 
 /**
@@ -20,7 +20,7 @@ public class Expense implements Serializable {
     @Id
     @Column(name = "ID")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     // 名前
     @Column(name = "NAME")
